Module 'sampling' renamed to 'sample'
