+++ kglib/kgcn/core/ingest/traverse/data/context_test.py
@@ -23,9 +23,9 @@ import grakn
 
 import kglib.kgcn.core.ingest.traverse.data.batch as batch
 import kglib.kgcn.core.ingest.traverse.data.neighbour as neighbour
-import kglib.kgcn.core.ingest.traverse.data.sampling.sampler as samp
+import kglib.kgcn.core.ingest.traverse.data.sample.sample as samp
 import kglib.kgcn.core.ingest.traverse.data.context as context
-import kglib.kgcn.core.ingest.traverse.data.sampling.ordered as ordered
+import kglib.kgcn.core.ingest.traverse.data.sample.ordered as ordered
 import kglib.kgcn.core.ingest.traverse.data.context_mocks as mocks
